Structure refactor (#82)
This PR updates the structure of the code:

- Acados becomes its own submodule to start decoupling from the Acados
code. Thus, the acados pytorch layer is also moved into that subfolder.
- We provide a new way how to print dataclass configs
- Tests are now properly split according to the submodules.
- We split the different utility functions into separate modules.

---------
