+++ leap_c/acados/layer.py
@@ -1,17 +1,13 @@
-"""Provides a PyTorch module that wraps a CasADi expression.
-
-TODO (Jasper): Does JIT compilation speed up the module?
-"""
 from dataclasses import dataclass
+from typing import Any
 
 import casadi as ca
 import numpy as np
 import torch
-import torch.autograd as autograd
-from acados_template import AcadosSimSolver
+from torch import autograd as autograd, nn as nn
 
-from leap_c.mpc import Mpc, MpcBatchedState, MpcInput, MpcParameter
-from leap_c.utils import tensor_to_numpy
+from acados_template import AcadosSimSolver
+from leap_c.acados.mpc import Mpc, MpcParameter, MpcBatchedState, MpcInput, MpcOutput
 
 
 class AutogradCasadiFunction(autograd.Function):
@@ -282,3 +278,115 @@ class MPCSolutionFunction(autograd.Function):
         # print("After: Grad p min", grad_p.min(), "Grad p max", grad_p.max())
 
         return (None, grad_x, grad_u, grad_p, None, None, None)
+
+
+def tensor_to_numpy(tensor: torch.Tensor):
+    return tensor.detach().cpu().numpy()
+
+
+class CasadiExprModule(nn.Module):
+    def __init__(self, expr: ca.SX, inputs: list[ca.SX]):
+        super().__init__()
+        self.expr = expr
+        self.inputs = inputs
+        self.f = ca.Function("f", inputs, [expr])
+
+        # setup backward AD
+        inputs_cat = ca.vertcat(*inputs)
+        grad_output = ca.SX.sym("output", expr.shape[0])
+        jtimes = ca.jtimes(expr, inputs_cat, grad_output, True)
+        self.jtimes = ca.Function("jtimes", [inputs_cat, grad_output], [jtimes])
+
+    def forward(self, *inputs: torch.Tensor) -> torch.Tensor:
+        assert len(inputs) == len(self.inputs)
+        return AutogradCasadiFunction.apply(self.f, self.jtimes, *inputs)
+
+
+class AcadosSimModule(nn.Module):
+    """A PyTorch module to wrap around AcadosSimSolver."""
+
+    def __init__(self, sim: AcadosSimSolver):
+        super().__init__()
+        self.sim = sim
+
+    def forward(
+        self, x: torch.Tensor, u: torch.Tensor, p: torch.Tensor | None = None
+    ) -> torch.Tensor:
+        return DynamicsSimFunction.apply(self.sim, x, u, p)
+
+
+class MpcSolutionModule(nn.Module):
+    """A PyTorch module to allow differentiating the solution given by an MPC planner,
+    with respect to some inputs.
+
+    Backpropagation works by using the sensitivities
+    given by the MPC. If differentiation with respect to parameters is desired, they must
+    be declared as global over the horizon (contrary to stagewise parameters).
+
+    NOTE: Make sure that you follow the documentation of AcadosOcpSolver.eval_solution_sensitivity
+        and AcadosOcpSolver.eval_and_get_optimal_value_gradient
+        or else the gradients used in the backpropagation might be erroneous!
+
+    NOTE: The status output can be used to rid yourself from non-converged solutions, e.g., by using the
+        CleanseAndReducePerSampleLoss module.
+
+    Attributes:
+        mpc: The MPC object to use.
+    """
+
+    def __init__(self, mpc: Mpc):
+        super().__init__()
+        self.mpc = mpc
+
+    def forward(
+        self,
+        mpc_input: MpcInput,
+        mpc_state: MpcBatchedState | None = None,
+    ) -> tuple[MpcOutput, MpcBatchedState, dict[str, Any]]:
+        """Differentiation is only allowed with respect to x0, u0 and p_global.
+
+        Args:
+            mpc_input: The MPCInput object containing the input that should be set.
+                NOTE x0, u0 and p_global must be tensors, if not None.
+            mpc_state: The MPCBatchedState object containing the initializations for the solver.
+
+        Returns:
+            mpc_output: An MPCOutput object containing tensors of u0, value (or Q, if u0 was given) and status of the solution.
+            mpc_state: The MPCBatchedState containing the iterates of the solution.
+            stats: A dictionary containing statistics from the MPC evaluation.
+        """
+        if mpc_input.parameters is None:
+            p_glob = None
+            p_rest = None
+        else:
+            p_glob = mpc_input.parameters.p_global
+            p_rest = mpc_input.parameters._replace(p_global=None)
+
+        ctx = ModuleCtx()
+        u0, value, status, state = MPCSolutionFunction.apply(  # type:ignore
+            self.mpc,
+            mpc_input.x0,
+            mpc_input.u0,
+            p_glob,
+            p_rest,
+            mpc_state,
+            ctx,
+        )
+
+        if ctx.dudp_global is not None:
+            dudp_global = torch.tensor(ctx.dudp_global, device=u0.device, dtype=u0.dtype)
+        else:
+            dudp_global = None
+
+        if mpc_input.u0 is None:
+            V = value
+            Q = None
+        else:
+            Q = value
+            V = None
+
+        return (
+            MpcOutput(u0=u0, Q=Q, V=V, status=status, du0_dp_global=dudp_global),
+            state,
+            self.mpc.last_call_stats,
+        )
