+++ leap_c/examples/quadrotor/mpc.py
@@ -7,8 +7,8 @@ import scipy
 from acados_template import AcadosOcp
 from leap_c.examples.quadrotor.casadi_models import get_rhs_quadrotor
 from leap_c.examples.quadrotor.utils import read_from_yaml
-from leap_c.mpc import Mpc
-from leap_c.utils import set_standard_sensitivity_options
+from leap_c.acados.mpc import Mpc
+from leap_c.acados.utils import set_standard_sensitivity_options
 
 PARAMS = OrderedDict(
     [
