+++ leap_c/run.py
@@ -1,16 +1,14 @@
 """Module for experiments and resurrecting trainers."""
 from argparse import ArgumentParser
-from dataclasses import asdict
 import datetime
 from pathlib import Path
-
-import yaml
 
 import leap_c.examples  # noqa: F401
 from leap_c.registry import create_default_cfg, create_task, create_trainer
 import leap_c.rl
 from leap_c.trainer import BaseConfig
-from leap_c.utils import log_git_hash_and_diff
+from leap_c.utils.cfg import cfg_as_python
+from leap_c.utils.git import log_git_hash_and_diff
 
 
 def print_inputs(
@@ -29,7 +27,8 @@ def print_inputs(
 
     # Report on the configuration
     print("\nConfiguration:")
-    print(yaml.dump(asdict(cfg), default_flow_style=False))
+    print(cfg_as_python(cfg))
+    print("\n")
 
 
 def default_output_path(trainer_name: str, task_name: str, seed: int) -> Path:
@@ -97,6 +96,7 @@ def create_parser() -> ArgumentParser:
     """
     parser = ArgumentParser()
     parser.add_argument("--output_path", type=Path, default=None)
+    parser.add_argument("-v", "--verbose", action="store_true")
     parser.add_argument("--trainer", type=str, default="sac_fop")
     parser.add_argument("--task", type=str, default="pendulum_swingup")
     parser.add_argument("--device", type=str, default="cpu")
@@ -113,6 +113,7 @@ if __name__ == "__main__":
     cfg = create_default_cfg(args.trainer)
     cfg.seed = args.seed
     cfg.train.steps = args.steps
+    cfg.log.verbose = args.verbose
 
     if args.output_path is None:
         output_path = default_output_path(args.trainer, args.task, cfg.seed)
