+++ leap_c/utils/seed.py
@@ -1,0 +1,13 @@
+import random
+
+import numpy as np
+import torch
+
+
+def set_seed(seed: int):
+    """Set the seed for all random number generators."""
+    random.seed(seed)
+    np.random.seed(seed)
+    torch.manual_seed(seed)
+    torch.backends.cudnn.deterministic = True
+    torch.backends.cudnn.benchmark = False
