+++ tests/leap_c/acados/test_layer.py
@@ -4,8 +4,8 @@ import torch
 from leap_c.examples.chain.mpc import ChainMpc
 from leap_c.examples.pendulum_on_a_cart.mpc import PendulumOnCartMPC
 from leap_c.examples.pointmass.mpc import PointMassMPC
-from leap_c.mpc import MpcInput, MpcParameter
-from leap_c.nn.modules import CleanseAndReducePerSampleLoss, MpcSolutionModule
+from leap_c.acados.mpc import MpcInput, MpcParameter
+from leap_c.acados.layer import MpcSolutionModule
 
 
 def test_MPCSolutionModule_on_PointMassMPC(
@@ -691,108 +691,3 @@ def test_MPCSolutionModule_on_ChainCost(
         return mpc_output.Q, mpc_output.status
 
     torch.autograd.gradcheck(only_dQdu0, u0, atol=1e-2, eps=1e-4, raise_exception=True)
-
-
-def test_CleanseAndReduce():
-    cleansed_loss = CleanseAndReducePerSampleLoss(
-        reduction="mean",
-        num_batch_dimensions=1,
-        n_nonconvergences_allowed=2,
-        throw_exception_if_exceeded=False,
-    )
-
-    x = torch.tensor([[1.0], [2.0], [3.0], [4.0]], dtype=torch.float64)
-    status = torch.tensor([[0], [1], [0], [0]], dtype=torch.int8)
-    loss, _ = cleansed_loss(x, status)
-    assert loss.item() == 8 / 3
-
-    status = torch.tensor([[0], [0], [0], [0]], dtype=torch.int8)
-    loss, _ = cleansed_loss(x, status)
-    assert loss.item() == 10 / 4
-
-    status = torch.tensor([[2], [0], [1], [0]], dtype=torch.int8)
-    loss, _ = cleansed_loss(x, status)
-    assert loss.item() == 6 / 2
-
-    status = torch.tensor([[2], [2], [1], [0]], dtype=torch.int8)
-    loss, _ = cleansed_loss(x, status)
-    assert loss.item() == 0.0
-
-    status = torch.tensor([[1], [0], [1]], dtype=torch.int8)
-    try:
-        loss, _ = cleansed_loss(x, status)
-        assert False
-    except ValueError:
-        assert True
-
-
-def test_CleanseAndReduceMultipleBatchAndSampleDims():
-    cleansed_loss = CleanseAndReducePerSampleLoss(
-        reduction="mean",
-        num_batch_dimensions=2,
-        n_nonconvergences_allowed=4,
-        throw_exception_if_exceeded=False,
-    )
-
-    x = torch.ones((3, 3, 3, 3))
-    x[0, 0] = 2
-    x[0, 1] = 100
-    status = torch.zeros((3, 3, 1), dtype=torch.int8)
-    status[0, 0] = 1
-    status[0, 1] = 2
-    loss, _ = cleansed_loss(x, status)
-    assert loss.item() == 1.0
-
-    x = torch.ones((3, 3, 3, 3))
-    status = torch.zeros((3, 3, 1), dtype=torch.int8)
-    status[0, 0] = 1
-    status[0, 1] = 2
-    status[0, 2] = 4
-    status[1, 2] = 5
-    loss, _ = cleansed_loss(x, status)
-    assert loss.item() == 1.0
-
-    status = torch.zeros((3, 3, 1), dtype=torch.int8)
-    status[0, 0] = 1
-    status[0, 1] = 2
-    status[0, 2] = 1
-    status[1, 2] = 2
-    status[2, 2] = 2
-    loss, _ = cleansed_loss(x, status)
-    assert loss.item() == 0.0
-
-    status = torch.zeros((3, 3, 3, 1), dtype=torch.int8)
-    try:
-        loss, _ = cleansed_loss(x, status)
-        assert False
-    except ValueError:
-        assert True
-
-    status = torch.zeros((3, 1), dtype=torch.int8)
-    try:
-        loss, _ = cleansed_loss(x, status)
-        assert False
-    except ValueError:
-        assert True
-
-    status = torch.ones((3, 3, 1), dtype=torch.int8)
-    status[0, 0] = 1
-    status[0, 1] = 2
-    status[0, 2] = 1
-    status[1, 2] = 2
-    status[2, 2] = 2
-    cleansed_loss = CleanseAndReducePerSampleLoss(
-        reduction="mean",
-        num_batch_dimensions=2,
-        n_nonconvergences_allowed=4,
-        throw_exception_if_exceeded=True,
-    )
-    try:
-        loss, _ = cleansed_loss(x, status)
-        assert False
-    except ValueError:
-        assert True
-
-
-if __name__ == "__main__":
-    pytest.main([__file__])
